+++ CHANGELOG.md
@@ -134,6 +134,7 @@ Semantic versioning in our case means:
 - Fixes several violations not been detected in `case:` statements
 - Fixes `WPS434` not detecting reassignments in `:=` case
 - Fixes `WPS314` not detecting `match` statements
+- Fixes `match` + `case` does not increase cognitive complexity
 
 ### Misc